requirements.txt
discord
disnake-pagination
discordpy-antispam
